Refactor

+++ src/main/java/org/example/simulation/logic/animalBehaviour/types/Stroll.java
@@ -37,13 +37,13 @@ public class Stroll implements Behaviour {
                 setNewRandomVelocity(animal);
             }
             numberOfTrials++;
-        } while(numberOfTrials < MAX_TRIALS_OF_DIRECTION_CHANGE);
+        } while (numberOfTrials < MAX_TRIALS_OF_DIRECTION_CHANGE);
     }
 
     private void setNewRandomVelocity(Animal animal) {
-        float nextVelocityX = Math.round(Math.random()*MAX_ANIMAL_SPEED * 2) - MAX_ANIMAL_SPEED;
+        float nextVelocityX = (float) Math.round(Math.random() * MAX_ANIMAL_SPEED * 2) - MAX_ANIMAL_SPEED;
         int randomSign = Math.random() < 0.5 ? 1 : -1;
-        float nextVelocityY = (float) Math.sqrt(Math.pow(MAX_ANIMAL_SPEED,2) - Math.pow(nextVelocityX,2)) * randomSign;
+        float nextVelocityY = (float) Math.sqrt(Math.pow(MAX_ANIMAL_SPEED, 2) - Math.pow(nextVelocityX, 2)) * randomSign;
         animal.setVelocityX(nextVelocityX);
         animal.setVelocityY(nextVelocityY);
     }
